% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVParser.R
\docType{data}
\name{CSVParser}
\alias{CSVParser}
\title{CSV File Parser}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CSVParser
}
\description{
Reads a .csv document
}
\author{
Jelmer Draaijer
}
\keyword{datasets}

