% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqualGainModel.R
\docType{data}
\name{EqualGainModel}
\alias{EqualGainModel}
\title{The main body of the model.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EqualGainModel
}
\description{
The main body of the model.
}
\author{
Jelmer Draaijer
}
\keyword{datasets}

