% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeHistory.R
\docType{data}
\name{ExchangeHistory}
\alias{ExchangeHistory}
\title{The history of an actor his exchanges}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExchangeHistory
}
\description{
The history of an actor his exchanges
}
\author{
Jelmer Draaijer
}
\keyword{datasets}

