% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangePair.R
\docType{data}
\name{ExchangePair}
\alias{ExchangePair}
\title{A set of two actors and two issues which can exchange}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExchangePair
}
\description{
A set of two actors and two issues which can exchange
}
\author{
Jelmer Draaijer
}
\keyword{datasets}

