% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBS.R
\docType{data}
\name{NBS}
\alias{NBS}
\title{Nash Barganing Solution}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NBS
}
\description{
Nash Barganing Solution
}
\author{
Jelmer Draaijer
}
\keyword{datasets}

